Figured out new "routing" scheme
Basically I don't want to use actual routes, because they caused a lot
of problems with hosting on gh-pages, which is the end-goal
here. Second, I don't think state for each route would be preserved if
going that route (unless I separate the state from the views ofc, but I
don't necessarily want to do that if they're only used within a specific
view, for example).

The current scheme is pretty simple; I don't need nested routes or
anything, state is preserved when navigating, and it's SILLY fast. I
think this was the best solution givin the context.